backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
idna==3.11
isort==7.0.0
oauthlib==3.3.1
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.2
